🧪 TQ70WD context: split raw deletion regression
