+++ packages/agentplane/src/commands/context/release-readiness.test.ts
@@ -297,60 +297,6 @@ describe("context release readiness guards", () => {
     ).toBe(true);
   });
 
-  it("keeps curated search usable after the raw corpus is removed", async () => {
-    const root = await tempRoot();
-    await write(root, "context/raw/specs/payment-api.md", "# Raw\n\nrawonlysentinel\n");
-    await write(
-      root,
-      "context/wiki/payment-api.md",
-      "# Payment API\n\nDurable curated payment contract sentinel.\n",
-    );
-    await write(
-      root,
-      ".agentplane/context/derived/facts/facts.jsonl",
-      JSON.stringify({
-        id: "fact.payment-api.durable",
-        summary: "Durable curated payment contract sentinel.",
-        source_refs: ["context/raw/specs/payment-api.md#L1-L3"],
-      }) + "\n",
-    );
-    await cmdContextReindex({
-      cwd: root,
-      parsed: { includeTasks: false, includeRaw: true, reset: true },
-    });
-    await rm(path.join(root, "context/raw/specs/payment-api.md"));
-    await cmdContextReindex({
-      cwd: root,
-      parsed: { includeTasks: false, includeRaw: false, reset: true },
-    });
-    const out = vi.spyOn(process.stdout, "write").mockImplementation(() => true);
-
-    await cmdContextSearch({
-      cwd: root,
-      parsed: {
-        query: "durable curated payment contract",
-        scope: "wiki,facts,graph,capabilities",
-        format: "json",
-        explain: false,
-      },
-    });
-    await cmdContextSearch({
-      cwd: root,
-      parsed: { query: "rawonlysentinel", scope: "raw", format: "json", explain: false },
-    });
-
-    const payloads = out.mock.calls
-      .map((call) => String(call[0]))
-      .join("")
-      .trim()
-      .split(/\n(?=\{)/u)
-      .map((chunk) => JSON.parse(chunk) as { results: { ref: string }[] });
-    expect(payloads[0]?.results.some((result) => result.ref.startsWith("context/wiki/"))).toBe(
-      true,
-    );
-    expect(payloads[1]?.results).toEqual([]);
-  });
-
   it("excludes private raw files from fallback search for raw and all scopes", async () => {
     const root = await tempRoot();
     await write(root, "context/raw/specs/payment-api.md", "# Payment API\n\nNeedle public.\n");
